web/requirements.txt
flask
splitwise
python-dotenv
python-dateutil
gunicorn